+++ src/checks/unordered_key.rs
@@ -9,21 +9,17 @@ pub(crate) struct UnorderedKeyChecker<'a> {
 
 impl UnorderedKeyChecker<'_> {
     fn message(&self, key_one: &str, key_two: &str) -> String {
-        return format!(
-            "{}: {}",
-            self.name,
-            self.template
-                .replace("{1}", key_one)
-                .replace("{2}", key_two)
-        );
+        self.template
+            .replace("{1}", key_one)
+            .replace("{2}", key_two)
     }
 }
 
 impl Default for UnorderedKeyChecker<'_> {
     fn default() -> Self {
         Self {
-            keys: Vec::new(),
             name: "UnorderedKey",
+            keys: Vec::new(),
             template: "The {1} key should go before the {2} key",
         }
     }
@@ -46,7 +42,7 @@ impl Check for UnorderedKeyChecker<'_> {
 
             let another_key = sorted_keys.get(index + 1)?;
 
-            let warning = Warning::new(line.clone(), self.message(&key, &another_key));
+            let warning = Warning::new(line.clone(), self.name(), self.message(&key, &another_key));
             return Some(warning);
         }
 
@@ -157,7 +153,8 @@ mod tests {
                         },
                         raw_string: String::from("BAR=FOO"),
                     },
-                    String::from("UnorderedKey: The BAR key should go before the FOO key"),
+                    "UnorderedKey",
+                    String::from("The BAR key should go before the FOO key"),
                 )),
             ),
         ];
@@ -200,7 +197,8 @@ mod tests {
                         },
                         raw_string: String::from("BAR=FOO"),
                     },
-                    String::from("UnorderedKey: The BAR key should go before the FOO key"),
+                    "UnorderedKey",
+                    String::from("The BAR key should go before the FOO key"),
                 )),
             ),
             (
@@ -223,7 +221,8 @@ mod tests {
                         },
                         raw_string: String::from("ABC=BAR"),
                     },
-                    String::from("UnorderedKey: The ABC key should go before the BAR key"),
+                    "UnorderedKey",
+                    String::from("The ABC key should go before the BAR key"),
                 )),
             ),
         ];
@@ -266,7 +265,8 @@ mod tests {
                         },
                         raw_string: String::from("BAR=FOO"),
                     },
-                    String::from("UnorderedKey: The BAR key should go before the FOO key"),
+                    "UnorderedKey",
+                    String::from("The BAR key should go before the FOO key"),
                 )),
             ),
             (
@@ -289,7 +289,8 @@ mod tests {
                         },
                         raw_string: String::from("DDD=BAR"),
                     },
-                    String::from("UnorderedKey: The DDD key should go before the FOO key"),
+                    "UnorderedKey",
+                    String::from("The DDD key should go before the FOO key"),
                 )),
             ),
         ];
@@ -332,7 +333,8 @@ mod tests {
                         },
                         raw_string: String::from("BAR=FOO"),
                     },
-                    String::from("UnorderedKey: The BAR key should go before the FOO key"),
+                    "UnorderedKey",
+                    String::from("The BAR key should go before the FOO key"),
                 )),
             ),
             (
@@ -355,7 +357,8 @@ mod tests {
                         },
                         raw_string: String::from("DDD=BAR"),
                     },
-                    String::from("UnorderedKey: The DDD key should go before the FOO key"),
+                    "UnorderedKey",
+                    String::from("The DDD key should go before the FOO key"),
                 )),
             ),
             (